migrate smartparser in tsx file
