+++ src/components/Overview/ParsingTest.tsx
@@ -1,0 +1,9 @@
+import { SmartParser } from '../../utils/SmartParser';
+
+// Create a random plot to use console.log()
+const EmptyDiv = () => {
+  return <></>;
+  const FunctionOutput = SmartParser;
+};
+
+export default EmptyDiv;
