+++ src/pages/Overview.tsx
@@ -5,7 +5,7 @@ import OverviewProfile from '../components/Overview/OverviewProfile';
 import TrackedChart from '../components/Overview/PlotlyCharts/TrackedChart';
 import DataCollectedChart from '../components/Overview/PlotlyCharts/DataCollectedChart';
 import DataTable from '../components/Overview/PlotlyCharts/DataTable';
-import EmptyDiv from '../components/Overview/Parsing';
+import EmptyDiv from '../components/Overview/ParsingTest';
 
 // MUI components
 import { Box, Grid } from '@mui/material';
